Move gamecontroller code to it's own demo
